remove files not related to TinyGarble2

+++ Benchmarks/aes/compile.sh
@@ -1,5 +1,16 @@
-#!/bin/bash
-mkdir -p syn
-design_vision -no_gui -f aes.dcsh
+echo "Use -d to compile with Synopsys Design Compiler"
 
-rm *.pvl *.syn *.mr *.log *.svf
+mkdir -p syn
+
+if [ $1 ] && [ $1 = "-d" ]; then
+	design_vision -no_gui -f aes.dcsh
+	rm *.pvl *.syn *.mr *.log *.svf
+else
+	yosys -c aes_1cc.tcl
+fi
+
+for verilogfile in syn/*.v
+do
+  empfile=${verilogfile%.*}.emp
+  ../../Verilog2EMP/bin/V2EMP_Main -i $verilogfile -o $empfile --log2std 
+done
